requirements.txt
crewai
langchain
crewai-tools
psycopg2-binary
python-dotenv
requests
schedule
langchain-openai
